Created class to customize all viewsets and applied on ipa app
