+++ ipa/views.py
@@ -1,4 +1,3 @@
-from rest_framework import viewsets, pagination, response
 from .serializers import ParticipantInstitutionSerializer
 from .serializers import ParticipantInstitutionTypeSerializer
 from .serializers import ContactSerializer
@@ -10,65 +9,51 @@ from .serializers import NoBreakSerializer
 from .models import NoBreak
 from .models import Switch
 from sigi_op.serializers import SiteTypeSerializer, SiteSerializer
+from sigi_op.views import CustomViewSet
 
 # Create your views here.
 
 
-class IpaListViewSet(viewsets.ModelViewSet):
+class IpaListViewSet(CustomViewSet):
     queryset = ParticipantInstitution.objects.all().order_by('name')
     serializer_class = ParticipantInstitutionSerializer
 
-    def list(self, request):
-        queryset = ParticipantInstitution.objects.all().order_by('name')
-        if request.GET.get('all'):
-            self.pagination_class = None
-            serializer = ParticipantInstitutionSerializer(queryset, many=True)
-            return response.Response(serializer.data)
-        else:
-            paginator = pagination.PageNumberPagination()
-            queryset = paginator.paginate_queryset(
-                queryset=queryset,
-                request=request
-                )
-            serializer = ParticipantInstitutionSerializer(queryset, many=True)
-            return paginator.get_paginated_response(serializer.data)
 
-
-class IpaTypeListViewSet(viewsets.ModelViewSet):
+class IpaTypeListViewSet(CustomViewSet):
     queryset = InstitutionType.objects.all().order_by('description')
     serializer_class = ParticipantInstitutionTypeSerializer
 
 
-class SiteListViewSet(viewsets.ModelViewSet):
+class SiteListViewSet(CustomViewSet):
     queryset = Site.objects.all().order_by('name')
     serializer_class = SiteSerializer
 
 
-class SiteTypeListViewSet(viewsets.ModelViewSet):
+class SiteTypeListViewSet(CustomViewSet):
     queryset = SiteType.objects.all().order_by('description')
     serializer_class = SiteTypeSerializer
 
 
-class ContactViewSet(viewsets.ModelViewSet):
+class ContactViewSet(CustomViewSet):
     queryset = Contact.objects.all().order_by('name')
     serializer_class = ContactSerializer
 
 
-class ContactTypeViewSet(viewsets.ModelViewSet):
+class ContactTypeViewSet(CustomViewSet):
     queryset = ContactType.objects.all().order_by('description')
     serializer_class = ContactTypeSerializer
 
 
-class GeneratorListViewSet(viewsets.ModelViewSet):
+class GeneratorListViewSet(CustomViewSet):
     queryset = Generator.objects.all().order_by('site_id')
     serializer_class = GeneratorSerializer
 
 
-class NoBreakViewSet(viewsets.ModelViewSet):
+class NoBreakViewSet(CustomViewSet):
     queryset = NoBreak.objects.all().order_by('site_id')
     serializer_class = NoBreakSerializer
 
 
-class SwitchViewSet(viewsets.ModelViewSet):
+class SwitchViewSet(CustomViewSet):
     queryset = Switch.objects.all().order_by('site_id')
     serializer_class = SwitchSerializer
