+++ sigi_op/views.py
@@ -1,7 +1,7 @@
 from django.contrib.auth import authenticate
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
-from rest_framework import status
+from rest_framework import status, viewsets, pagination
 from rest_framework.authtoken.models import Token
 from .serializers import UserSerializer
 from django.contrib.auth.models import User
@@ -36,3 +36,20 @@ def login(request):
 
     token, _ = Token.objects.get_or_create(user=user)
     return Response({"username": username, "token": token.key})
+
+
+class CustomViewSet(viewsets.ModelViewSet):
+    def list(self, request):
+        queryset = self.queryset
+        if request.GET.get('all'):
+            self.pagination_class = None
+            serializer = self.serializer_class(queryset, many=True)
+            return Response(serializer.data)
+        else:
+            paginator = pagination.PageNumberPagination()
+            queryset = paginator.paginate_queryset(
+                queryset=queryset,
+                request=request
+                )
+            serializer = self.serializer_class(queryset, many=True)
+            return paginator.get_paginated_response(serializer.data)
